+++ source/profile.h
@@ -11,7 +11,7 @@ extern "C" {
 // be declared read-only (by hiding the variable). But this is not c++, and
 // I need to access this variable very quickly for the timers (see timer.c)
 // If you want to change this variable, do it in the load_game function only !
-extern int fps;
+extern float fps;
 
 void switch_fps_mode(void);
 void init_fps(void);
